go: consolidate procman.go into main.go
Single-file structure is idiomatic for small CLI tools that aren't
intended as libraries. At ~320 lines, the code is easily scannable
in one file without jumping between files.

This also improves consistency with mdembed, another pinned repo
with similar single-purpose CLI tool characteristics.

Ordering: imports → const/types → main() → functions by receiver.
